+++ CHANGELOG.md
@@ -6,6 +6,7 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 
 Features:
 - Add modifier `k-buttonIcon--micro` on `k-ButtonIcon`.
+- Add `initialInstallment` option to `LoanSimulator`.
 
 Fixes:
 - Move `babel-preset-airbnb` module from `devDependencies` to `dependencies`.